[Refactor] Split Dashboard component into DashboardContent and DashboardActions for improved structure and maintainability
